Results of Auckland Local Elections 2022

Results of Auckland Local Elections 2022

Usage

auckland2022

Format

A data frame with 1076 observations and 13 variables. These are the vote totals and rank-ordering of candidates in the progress and final results of STV elections in Auckland, New Zealand, in 2022. "Progress results include votes returned up until Friday 7 October. Special votes and votes hand delivered by 12 noon on Saturday 8 October are not included... Final count includes all votes, including special votes that are returned by 12 noon on Saturday 8 October." We have calculated nBallots from the votes as reported by the Electoral Officer, see data-raw/auckland2022.R for details. Curiously, the reported votes are all integral, suggesting that some method other than Meek's was used to count the ballots, or that some postprocessing of the usual reporting of Meek's results was performed to avoid the (apparent) double-counting of votes which occurs whenever an excluded candidate's votes are both reported against this candidate and also distributed among the candidates still in play.