plot() method for the result of an experiment with varying numbers of ballots

The "adjusted rank" of a candidate is their ranking \(r\) plus their adjusted and scaled "winning margin". The scaled margin is \(e^{-cx/\sqrt{n}}\), where \(x\) is the adjusted margin (i.e. the number of votes by which this candidate is ahead of the next-weaker candidate, adjusted for the number of ballots \(n\) and the number of seats \(s\)), and \(c>0\) is the margin-scaling parameter cmargin. The adjusted margin \(x\) is computed from the unadjusted winning margin \(v\) by \(x = v(1+bs)/(1+a\sqrt{n})\) where \(a\) is the value of anBallots and \(b\) is the value of bnSeats. Note that the margin is unadjusted if \(a=0\) and \(b=0\).

Usage

# S3 method for SafeRankExpt
plot(
  x,
  facetWrap = FALSE,
  anBallots = 0,
  bnSeats = 0,
  cMargin = 1,
  xlab = "Ballots",
  ylab = "Adjusted Rank",
  main = NULL,
  line = TRUE,
  pointSize = 1,
  ...
)

Arguments

x

object containing experimental results

facetWrap

TRUE provides per-candidate scatterplots

anBallots, bnSeats, cMargin

parameters in the rank-adjustment formula

xlab, ylab

axis labels

main

overall title for the plot

line

TRUE will connect points with lines, and will disable jitter

pointSize

diameter of points

...

params for generic plot()

Value

graphics object, with side-effect in RStudio Plots pane

Details

A very small margin of victory causes a candidate to have an adjusted rank very near to \(r+1\).

The default value of cmargin=1.0 draws visual attention to candidates with a very small winning margin.

Setting cmargin=0 is a special case, causing the winning margin to be scaled in proportion to the number of seats and to \(1/\sqrt{n}\). This seems an appropriate scaling for STV elections, because the margin of victory for an elected candidate is typically not much larger than the quota of \(n/(s+1)\) (Droop) or \(n/s\) (Hare). The variance on a winning margin seems likely to be proportional to \(n\) in any real-world election. We suggest that the SafeVote package could aid election authorities in testing and refining this hypothesis, and then using it to determine whether a preliminary result from counting a fraction of the ballots is sufficiently "safe" to be announced to the public.

If you wish to see only ranks on the plot, use anBallots=0, bnSeats=0, cmargin=0.

If you wish to see unscaled margins, use anBallots=1, bnSeats=1, cmargin=0.

Todo: Accept a list of SafeVoteExpt objects.

Todo: Multiple counts with the same number of ballots could be summarised with a box-and-whisker graphic, rather than a set of jittered points.

Todo: Consider developing a non-exponential scaling that is appropriate for plotting stochastic experimental data derived from Condorcet elections.