Count preferential ballots using an STV method

Count preferential ballots using an STV method

stv(
  votes,
  nseats = NULL,
  eps = 0.001,
  equal.ranking = FALSE,
  fsep = "\t",
  ties = c("f", "b"),
  constant.quota = FALSE,
  quota.hare = FALSE,
  group.nseats = NULL,
  group.members = NULL,
  complete.ranking = FALSE,
  invalid.partial = FALSE,
  verbose = FALSE,
  seed = 1234,
  quiet = FALSE,
  digits = 3,
  safety = 1,
  ...
)

Arguments

votes

an array with one column per candidate and one row per ballot

nseats

the number of seats to be filled in this election

eps

fuzz-factor when comparing fractional votes

equal.ranking

TRUE if ballots are allowed to rank candidates equally

fsep

column-separator for output

ties

vector of tie-breaking methods: 'f' for forward, 'b' for backward

constant.quota

TRUE if quota is held constant. Over-rides quota.hare.

quota.hare

TRUE if Hare quota, FALSE if Droop quota

group.nseats

number of seats reserved to members of a group, as in the Church of England's STV methodology.

group.members

vector of members of a group with reserved seats

complete.ranking

TRUE if the balloting is intended to produce a ranking of all candidates. This affects the value assigned to nseats when stv() is called with nseats=NULL

invalid.partial

TRUE if ballots are invalid (aka "informal") if they do not specify a complete ranking of candidates

verbose

TRUE for diagnostic output

seed

integer seed for the (default) RNG in this instance of R

quiet

TRUE to suppress console output

digits

number of significant digits in the output table

safety

number of standard deviations on vote-counts, when producing a SafeRank by clustering near-ties in a complete ranking

...

undocumented

Value

object of class vote.stv

Examples

data(food_election)
stv(food_election, complete.ranking=TRUE, safety=0.5)
#> 
#> Results of Single transferable vote
#> ===================================                           
#> Number of valid votes:   20
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          5
#> 
#> Fuzz on SafeRank:	 2.236068
#> Complete ranking:
#> 
#> | Rank|Margin    |  Candidate   | Elected|SafeRank |
#> |----:|:---------|:------------:|-------:|:--------|
#> |    1|8.0000000 |  Chocolate   |       x|1        |
#> |    2|2.7771111 | Strawberries |       x|2        |
#> |    3|0.1114444 |   Oranges    |       x|3        |
#> |    4|1.8885556 |    Sweets    |       x|3        |
#> |    5|NA        |    Pears     |       x|3        |
#> 
#> Vote transfers:
#> 
#> |             |         1| 2-trans|            2| 3-trans|       3| 4-trans|      4| 5-trans|     5|
#> |:------------|---------:|-------:|------------:|-------:|-------:|-------:|------:|-------:|-----:|
#> |Quota        |     3.334|        |        3.334|        |   2.473|        |  1.964|        | 1.001|
#> |Oranges      |     4.000|   0.000|        4.000|   0.000|   4.000|  -1.527|       |        |      |
#> |Pears        |     2.000|   0.000|        2.000|   0.000|   2.000|   0.000|  2.000|   0.000| 2.000|
#> |Chocolate    |    12.000|  -8.666|             |        |        |        |       |        |      |
#> |Strawberries |     1.000|   5.777|        6.777|  -3.443|        |        |       |        |      |
#> |Sweets       |     1.000|   2.889|        3.889|   0.000|   3.889|   0.000|  3.889|  -1.925|      |
#> |Elected      | Chocolate|        | Strawberries|        | Oranges|        | Sweets|        | Pears|
#> |Eliminated   |          |        |             |        |        |        |       |        |      |
#> 
#> Elected: Chocolate, Strawberries, Oranges, Sweets, Pears 
#> 
stv(food_election, nseats = NULL, eps = 0.001,
  equal.ranking = FALSE, fsep = "\t", ties = c("f", "b"), 
  constant.quota = FALSE, quota.hare = FALSE, 
  group.nseats = NULL, group.members = NULL,
  complete.ranking = FALSE, invalid.partial = FALSE, 
  verbose = FALSE, seed = 1234, quiet = FALSE, digits = 3)
#> 
#> Results of Single transferable vote
#> ===================================                           
#> Number of valid votes:   20
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          2
#> 
#> Fuzz on SafeRank:	 4.472136
#> Complete ranking:
#> 
#> | Rank|Margin    |  Candidate   | Elected|SafeRank |
#> |----:|:---------|:------------:|-------:|:--------|
#> |    1|8.0000000 |  Chocolate   |       x|1        |
#> |    2|1.4451111 |   Oranges    |       x|2        |
#> |    3|1.7774444 | Strawberries |        |2        |
#> |    4|0.7774444 |    Sweets    |        |2        |
#> |    5|NA        |    Pears     |        |2        |
#> 
#> Vote transfers:
#> 
#> |             |         1| 2-trans|     2| 3-trans|      3| 4-trans|       4|
#> |:------------|---------:|-------:|-----:|-------:|------:|-------:|-------:|
#> |Quota        |     6.668|        | 6.667|        |  6.667|        |   5.278|
#> |Oranges      |     4.000|   0.000| 4.000|       2|  6.000|   0.000|   6.000|
#> |Pears        |     2.000|   0.000| 2.000|      -2|       |        |        |
#> |Chocolate    |    12.000|  -5.332|      |        |       |        |        |
#> |Strawberries |     1.000|   3.555| 4.555|       0|  4.555|   0.000|   4.555|
#> |Sweets       |     1.000|   1.777| 2.777|       0|  2.777|  -2.777|        |
#> |Elected      | Chocolate|        |      |        |       |        | Oranges|
#> |Eliminated   |          |        | Pears|        | Sweets|        |        |
#> 
#> Elected: Chocolate, Oranges 
#>