Test the sensitivity of a result to tactical voting.

Ballots are added until the ranking changes, or until a specified number of ballots (`alimit`) have been deleted. A simple tactic of "plumping" is used, if a `favoured` candidate is specified. Alternatively, a `tacticalBallot` may be specified.

test.additions(
  votes,
  alimit = NULL,
  ainc = NULL,
  favoured = NULL,
  tacticalBallot = NULL,
  rankMethod = "SafeRank",
  countMethod = "stv",
  countArgs = NULL
)

Arguments

votes

A set of ballots

alimit

Maximum number of plumping ballots to add

ainc

Number of ballots to be added in each step

favoured

Name of the candidate being "plumped". If NULL, a random candidate is selected from among the candidates not initially top-ranked. All other candidates are ranked #2. An integer value for 'favoured' is interpreted as an index into the candidate names.

tacticalBallot

A ballot paper i.e. a vector of length ncol(ballots). If this argument is non-null, it takes precedence over 'favoured' when the ballot box is being stuffed.

rankMethod

Name of a ranking attribute in the output of countMethod

countMethod

The name of a function which will count the ballots

countArgs

List of args to be passed to countMethod (in addition to votes)

Value

Number of plumping ballots required for the tactical ballot (if any) to affect the ranking, or for the favoured candidate to move up in the ranking; or zero, if the termination condition has been reached without the tactical voting (if any) having affected the ranking or the ranking of the favoured candidate (if any) having improved. The results of the final ballot count are attached as an attribute "finalResults".

Examples

{
data(food_election)
test.additions(food_election, countMethod="condorcet", alimit=10)
test.additions(food_election, favoured=1, countArgs=list(nseats=10))
test.additions(food_election, tacticalBallot=c(1,2,3,4,5), alimit=5)
}
#> Initial ranking:
#>      Oranges        Pears    Chocolate Strawberries       Sweets 
#>            2            2            1            2            2 
#> 
#> Adding up to 10 ballots to favour Sweets 
#> Testing progress:  3,
#> Results of Condorcet voting
#> ===========================                           
#> Number of valid votes:   23
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          1
#> 
#> 
#> |             | Oranges| Pears| Chocolate| Strawberries| Sweets| Total| Score | BordaRank | SafeRank| Winner| Loser|
#> |:------------|-------:|-----:|---------:|------------:|------:|-----:|:-----:|:---------:|--------:|------:|-----:|
#> |Oranges      |       0|     1|         0|            0|      0|     1|  22   |     4     |        2|       |      |
#> |Pears        |       0|     0|         0|            0|      0|     0|   8   |     5     |        5|       |     x|
#> |Chocolate    |       1|     1|         0|            1|      1|     4|  48   |     1     |        1|      x|      |
#> |Strawberries |       1|     1|         0|            0|      1|     3|  28   |     2     |        2|       |      |
#> |Sweets       |       1|     1|         0|            0|      0|     2|  28   |     2     |        2|       |      |
#> 
#> SafeRank fuzz on Borda scores: 9.591663
#> Gaps in Borda scores: min 0 mean 9.6 max 20 ; all 0 6 8 14 20
#> Condorcet winner: Chocolate
#> Condorcet loser: Pears
#> 
#> 
#> Ranking changed after 3 tactical ballots were stuffed.
#> Final ranking:
#>      Oranges        Pears    Chocolate Strawberries       Sweets 
#>            2            5            1            2            2 
#> Warning: Adjusting nseats (originally 10) to be in 1..5
#> Initial ranking:
#>    Chocolate Strawberries      Oranges       Sweets        Pears 
#>            1            2            2            2            2 
#> 
#> Adding up to 1 ballots to favour Oranges 
#> Warning: 
#> Orangesis already top-ranked.
#> Testing progress:  1,
#> Warning: Adjusting nseats (originally 10) to be in 1..5
#> Detected  1 invalid votes. Number of valid votes is 20 .
#> Use invalid.votes(...) function to view discarded records.
#> 
#> Results of Single transferable vote
#> ===================================                           
#> Number of valid votes:   20
#> Number of invalid votes:  1
#> Number of candidates:     5
#> Number of seats:          5
#> 
#> Fuzz on SafeRank:	 4.472136
#> Complete ranking:
#> 
#> | Rank|Margin    |  Candidate   | Elected|SafeRank |
#> |----:|:---------|:------------:|-------:|:--------|
#> |    1|8.0000000 |  Chocolate   |       x|1        |
#> |    2|2.7771111 | Strawberries |       x|2        |
#> |    3|0.1114444 |   Oranges    |       x|2        |
#> |    4|1.8885556 |    Sweets    |       x|2        |
#> |    5|NA        |    Pears     |       x|2        |
#> 
#> Vote transfers:
#> 
#> |             |         1| 2-trans|            2| 3-trans|       3| 4-trans|      4| 5-trans|     5|
#> |:------------|---------:|-------:|------------:|-------:|-------:|-------:|------:|-------:|-----:|
#> |Quota        |     3.334|        |        3.334|        |   2.473|        |  1.964|        | 1.001|
#> |Oranges      |     4.000|   0.000|        4.000|   0.000|   4.000|  -1.527|       |        |      |
#> |Pears        |     2.000|   0.000|        2.000|   0.000|   2.000|   0.000|  2.000|   0.000| 2.000|
#> |Chocolate    |    12.000|  -8.666|             |        |        |        |       |        |      |
#> |Strawberries |     1.000|   5.777|        6.777|  -3.443|        |        |       |        |      |
#> |Sweets       |     1.000|   2.889|        3.889|   0.000|   3.889|   0.000|  3.889|  -1.925|      |
#> |Elected      | Chocolate|        | Strawberries|        | Oranges|        | Sweets|        | Pears|
#> |Eliminated   |          |        |             |        |        |        |       |        |      |
#> 
#> Elected: Chocolate, Strawberries, Oranges, Sweets, Pears 
#> 
#> 
#> No improvement in ranking after 1 ballots were stuffed.
#> Initial ranking:
#>    Chocolate      Oranges Strawberries       Sweets        Pears 
#>            1            2            2            2            2 
#> 
#> Adding up to 5 ballots with preferences 1 2 3 4 5 
#> Testing progress:  2,
#> Results of Single transferable vote
#> ===================================                           
#> Number of valid votes:   22
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          2
#> 
#> Fuzz on SafeRank:	 4.690416
#> Complete ranking:
#> 
#> | Rank|Margin    |  Candidate   | Elected|SafeRank |
#> |----:|:---------|:------------:|-------:|:--------|
#> |    1|6.0000000 |  Chocolate   |       x|1        |
#> |    2|3.8895556 |   Oranges    |       x|2        |
#> |    3|1.5552222 | Strawberries |        |2        |
#> |    4|0.5552222 |    Sweets    |        |2        |
#> |    5|NA        |    Pears     |        |2        |
#> 
#> Vote transfers:
#> 
#> |             |         1| 2-trans|     2| 3-trans|       3|
#> |:------------|---------:|-------:|-----:|-------:|-------:|
#> |Quota        |     7.334|        | 7.334|        |   7.334|
#> |Oranges      |     6.000|   0.000| 6.000|       2|   8.000|
#> |Pears        |     2.000|   0.000| 2.000|      -2|        |
#> |Chocolate    |    12.000|  -4.666|      |        |        |
#> |Strawberries |     1.000|   3.110| 4.110|       0|   4.110|
#> |Sweets       |     1.000|   1.555| 2.555|       0|   2.555|
#> |Elected      | Chocolate|        |      |        | Oranges|
#> |Eliminated   |          |        | Pears|        |        |
#> 
#> Elected: Chocolate, Oranges 
#> 
#>  4,
#> Results of Single transferable vote
#> ===================================                           
#> Number of valid votes:   24
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          2
#> 
#> Fuzz on SafeRank:	 4.898979
#> Complete ranking:
#> 
#> | Rank|Margin |  Candidate   | Elected|SafeRank |
#> |----:|:------|:------------:|-------:|:--------|
#> |    1|4.000  |  Chocolate   |       x|1        |
#> |    2|6.334  |   Oranges    |       x|1        |
#> |    3|1.333  | Strawberries |        |3        |
#> |    4|0.333  |    Sweets    |        |3        |
#> |    5|NA     |    Pears     |        |3        |
#> 
#> Vote transfers:
#> 
#> |             |         1| 2-trans|     2| 3-trans|       3|
#> |:------------|---------:|-------:|-----:|-------:|-------:|
#> |Quota        |     8.001|        | 8.001|        |   8.001|
#> |Oranges      |     8.000|   0.000| 8.000|       2|  10.000|
#> |Pears        |     2.000|   0.000| 2.000|      -2|        |
#> |Chocolate    |    12.000|  -3.999|      |        |        |
#> |Strawberries |     1.000|   2.666| 3.666|       0|   3.666|
#> |Sweets       |     1.000|   1.333| 2.333|       0|   2.333|
#> |Elected      | Chocolate|        |      |        | Oranges|
#> |Eliminated   |          |        | Pears|        |        |
#> 
#> Elected: Chocolate, Oranges 
#> 
#> 
#> Ranking changed after 4 tactical ballots were stuffed.
#> Final ranking:
#>    Chocolate      Oranges Strawberries       Sweets        Pears 
#>            1            1            3            3            3