Test the sensitivity of a ranking to random deletions

Ballots are deleted until the ranking changes, or until a specified number of ballots (`dlimit`) have been deleted.

test.deletions(
  votes = "food_election",
  countMethod = "condorcet",
  countArgs = NULL,
  dlimit = NULL,
  dinc = NULL,
  rankMethod = "SafeRank"
)

Arguments

votes

A set of ballots

countMethod

The name of a function which will count the ballots

countArgs

List of args to be passed to countMethod (in addition to votes)

dlimit

Maximum number of ballots to delete

dinc

Number of ballots to be deleted in each step

rankMethod

Name of a ranking attribute in the output of countMethod

Value

A vector of the indices of ballots whose deletions affected the ranking. The vector is of zero length if the 'dlimit' termination condition had been reached without affecting the ranking.

Examples

{
data(food_election)
test.deletions(food_election, countMethod="condorcet", dlimit=4)
test.deletions(food_election, countMethod="stv", countArgs=list(nseats=10))
}
#> Initial ranking:
#>      Oranges        Pears    Chocolate Strawberries       Sweets 
#>            2            2            1            2            2 
#> 
#> Deleting up to 4 ballots at random:
#>  Testing progress: *
#> Results of Condorcet voting
#> ===========================                           
#> Number of valid votes:   18
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          1
#> 
#> 
#> |             | Oranges| Pears| Chocolate| Strawberries| Sweets| Total| Score | BordaRank | SafeRank| Winner| Loser|
#> |:------------|-------:|-----:|---------:|------------:|------:|-----:|:-----:|:---------:|--------:|------:|-----:|
#> |Oranges      |       0|     1|         0|            0|      1|     2|  22   |     2     |        2|       |      |
#> |Pears        |       0|     0|         0|            0|      0|     0|   8   |     5     |        2|       |     x|
#> |Chocolate    |       1|     1|         0|            1|      1|     4|  44   |     1     |        1|      x|      |
#> |Strawberries |       1|     1|         0|            0|      1|     3|  21   |     3     |        2|       |      |
#> |Sweets       |       0|     1|         0|            0|      0|     1|  16   |     4     |        2|       |      |
#> 
#> SafeRank fuzz on Borda scores: 8.485281
#> Gaps in Borda scores: min 1 mean 8.8 max 22 ; all 1 5 8 8 22
#> Condorcet winner: Chocolate
#> Condorcet loser: Pears
#> 
#> *
#> Results of Condorcet voting
#> ===========================                           
#> Number of valid votes:   16
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          1
#> 
#> 
#> |             | Oranges| Pears| Chocolate| Strawberries| Sweets| Total| Score | BordaRank | SafeRank| Winner| Loser|
#> |:------------|-------:|-----:|---------:|------------:|------:|-----:|:-----:|:---------:|--------:|------:|-----:|
#> |Oranges      |       0|     1|         0|            0|      1|     2|  22   |     2     |        2|       |      |
#> |Pears        |       0|     0|         0|            0|      0|     0|   8   |     5     |        4|       |     x|
#> |Chocolate    |       1|     1|         0|            1|      1|     4|  36   |     1     |        1|      x|      |
#> |Strawberries |       1|     1|         0|            0|      1|     3|  21   |     3     |        2|       |      |
#> |Sweets       |       0|     1|         0|            0|      0|     1|  10   |     4     |        4|       |      |
#> 
#> SafeRank fuzz on Borda scores: 8
#> Gaps in Borda scores: min 1 mean 7.2 max 14 ; all 1 2 8 11 14
#> Condorcet winner: Chocolate
#> Condorcet loser: Pears
#> 
#> 
#> Ranking changed after 4 ballots were deleted.
#> Final ranking:
#>      Oranges        Pears    Chocolate Strawberries       Sweets 
#>            2            4            1            2            4 
#> 
#> Safety violation after 4 ballot deletions.
#> 
#> Deleted ballots: 5 12 15 9 
#> Warning: Adjusting nseats (originally 10) to be in 1..5
#> Initial ranking:
#>    Chocolate Strawberries      Oranges       Sweets        Pears 
#>            1            2            2            2            2 
#> 
#> Deleting up to 1 ballots at random:
#>  Testing progress:  16,
#> Warning: Adjusting nseats (originally 10) to be in 1..5
#> 
#> Results of Single transferable vote
#> ===================================                           
#> Number of valid votes:   19
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          5
#> 
#> Fuzz on SafeRank:	 4.358899
#> Complete ranking:
#> 
#> | Rank|Margin    |  Candidate   | Elected|SafeRank |
#> |----:|:---------|:------------:|-------:|:--------|
#> |    1|9.0000000 |  Chocolate   |       x|1        |
#> |    2|2.9441111 | Strawberries |       x|2        |
#> |    3|0.9441111 |    Sweets    |       x|2        |
#> |    4|1.0000000 |   Oranges    |       x|2        |
#> |    5|NA        |    Pears     |       x|2        |
#> 
#> Vote transfers:
#> 
#> |             |         1| 2-trans|            2| 3-trans|      3| 4-trans|       4| 5-trans|     5|
#> |:------------|---------:|-------:|------------:|-------:|------:|-------:|-------:|-------:|-----:|
#> |Quota        |     3.168|        |        3.167|        |  2.237|        |   1.668|        | 1.001|
#> |Oranges      |     3.000|   0.000|        3.000|   0.000|  3.000|   0.000|   3.000|  -1.332|      |
#> |Pears        |     2.000|   0.000|        2.000|   0.000|  2.000|   0.000|   2.000|   0.000| 2.000|
#> |Chocolate    |    12.000|  -8.832|             |        |       |        |        |        |      |
#> |Strawberries |     1.000|   5.888|        6.888|  -3.721|       |        |        |        |      |
#> |Sweets       |     1.000|   2.944|        3.944|   0.000|  3.944|  -1.707|        |        |      |
#> |Elected      | Chocolate|        | Strawberries|        | Sweets|        | Oranges|        | Pears|
#> |Eliminated   |          |        |             |        |       |        |        |        |      |
#> 
#> Elected: Chocolate, Strawberries, Sweets, Oranges, Pears 
#> 
#> 
#> Ranking changed after 1 ballots were deleted.
#> Final ranking:
#>    Chocolate Strawberries       Sweets      Oranges        Pears 
#>            1            2            2            2            2 
#> 
#> Safety violation after 1 ballot deletions.
#> 
#> Deleted ballots: 16