Test the fraction of ballots required to determine a result.

Starting from some number ('astart') of randomly-selected ballots, additional randomly-selected ballots are added until the same result is obtained as when all of the ballots are counted.

test.fraction(
  votes,
  astart = NULL,
  ainc = NULL,
  arep = NULL,
  rankMethod = "SafeRank",
  countMethod = "stv",
  countArgs = NULL
)

Arguments

votes

A set of ballots

astart

Starting number of ballots (min 2)

ainc

Number of ballots to be added in each step

arep

Limit on the number of repetitions of the test. Required to be non-null if ainc==0.

rankMethod

Name of a ranking attribute in the output of countMethod

countMethod

The name of a function which will count the ballots

countArgs

List of args to be passed to countMethod (in addition to votes)

Value

Number of ballots required to reach the same result as when all the ballots are counted. The final ballot count is attached as an attribute "finalResults".

Examples

data(food_election)
test.fraction(food_election, countMethod="condorcet", 
              countArgs=list(safety=0.5))
#> Initial ranking (counting all ballots):
#>      Oranges        Pears    Chocolate Strawberries       Sweets 
#>            3            5            1            2            4 
#> 
#> Selecting an increasingly-large fraction of 20 ballots until the initial ranking is found.
#> Progress:  10%,
#> Results of Condorcet voting
#> ===========================                          
#> Number of valid votes:   2
#> Number of invalid votes: 0
#> Number of candidates:    5
#> Number of seats:         1
#> 
#> 
#> |             | Oranges| Pears| Chocolate| Strawberries| Sweets| Total| Score | BordaRank| SafeRank| Winner|
#> |:------------|-------:|-----:|---------:|------------:|------:|-----:|:-----:|---------:|--------:|------:|
#> |Oranges      |       0|     0|         0|            0|      0|     0|   0   |         4|        4|       |
#> |Pears        |       0|     0|         0|            0|      0|     0|   0   |         4|        4|       |
#> |Chocolate    |       1|     1|         0|            1|      1|     4|   8   |         1|        1|      x|
#> |Strawberries |       1|     1|         0|            0|      0|     2|   3   |         2|        2|       |
#> |Sweets       |       1|     1|         0|            0|      0|     2|   3   |         2|        2|       |
#> 
#> SafeRank fuzz on Borda scores: 1.414214
#> Gaps in Borda scores: min 0 mean 1.6 max 5 ; all 0 0 0 3 5
#> Condorcet winner: Chocolate
#> There is no condorcet loser (no candidate lost to all other candidates).
#> 
#>  30%,
#> Results of Condorcet voting
#> ===========================                          
#> Number of valid votes:   6
#> Number of invalid votes: 0
#> Number of candidates:    5
#> Number of seats:         0
#> 
#> 
#> |             | Oranges| Pears| Chocolate| Strawberries| Sweets| Total| Score | BordaRank| SafeRank| Loser|
#> |:------------|-------:|-----:|---------:|------------:|------:|-----:|:-----:|---------:|--------:|-----:|
#> |Oranges      |       0|     1|         1|            0|      1|     3|  11   |         1|        1|      |
#> |Pears        |       0|     0|         0|            0|      1|     1|   4   |         4|        4|      |
#> |Chocolate    |       0|     1|         0|            1|      1|     3|   8   |         3|        1|      |
#> |Strawberries |       0|     1|         0|            0|      1|     2|  10   |         2|        1|      |
#> |Sweets       |       0|     0|         0|            0|      0|     0|   0   |         5|        5|     x|
#> 
#> SafeRank fuzz on Borda scores: 2.44949
#> Gaps in Borda scores: min 0 mean 2.2 max 4 ; all 0 1 2 4 4
#> There is no condorcet winner (no candidate won over all other candidates).
#> Condorcet loser: Sweets
#> 
#>  50%,
#> Results of Condorcet voting
#> ===========================                           
#> Number of valid votes:   10
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          1
#> 
#> 
#> |             | Oranges| Pears| Chocolate| Strawberries| Sweets| Total| Score | BordaRank| SafeRank| Winner|
#> |:------------|-------:|-----:|---------:|------------:|------:|-----:|:-----:|---------:|--------:|------:|
#> |Oranges      |       0|     1|         1|            1|      1|     4|  18   |         1|        1|      x|
#> |Pears        |       0|     0|         0|            0|      0|     0|   8   |         4|        3|       |
#> |Chocolate    |       0|     1|         0|            1|      1|     3|  16   |         2|        1|       |
#> |Strawberries |       0|     0|         0|            0|      0|     0|   6   |         5|        3|       |
#> |Sweets       |       0|     1|         0|            1|      0|     2|  10   |         3|        3|       |
#> 
#> SafeRank fuzz on Borda scores: 3.162278
#> Gaps in Borda scores: min 2 mean 3.6 max 6 ; all 2 2 2 6 6
#> Condorcet winner: Oranges
#> There is no condorcet loser (no candidate lost to all other candidates).
#> 
#>  70%,
#> Results of Condorcet voting
#> ===========================                           
#> Number of valid votes:   14
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          1
#> 
#> 
#> |             | Oranges| Pears| Chocolate| Strawberries| Sweets| Total| Score | BordaRank| SafeRank| Winner|
#> |:------------|-------:|-----:|---------:|------------:|------:|-----:|:-----:|---------:|--------:|------:|
#> |Oranges      |       0|     1|         0|            0|      1|     2|  18   |         3|        3|       |
#> |Pears        |       0|     0|         0|            0|      0|     0|   8   |         4|        4|       |
#> |Chocolate    |       1|     1|         0|            1|      1|     4|  28   |         1|        1|      x|
#> |Strawberries |       1|     1|         0|            0|      1|     3|  22   |         2|        2|       |
#> |Sweets       |       0|     0|         0|            0|      0|     0|   7   |         5|        4|       |
#> 
#> SafeRank fuzz on Borda scores: 3.741657
#> Gaps in Borda scores: min 1 mean 5.6 max 10 ; all 1 4 6 7 10
#> Condorcet winner: Chocolate
#> There is no condorcet loser (no candidate lost to all other candidates).
#> 
#>  90%,
#> Results of Condorcet voting
#> ===========================                           
#> Number of valid votes:   18
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          1
#> 
#> 
#> |             | Oranges| Pears| Chocolate| Strawberries| Sweets| Total| Score | BordaRank | SafeRank| Winner| Loser|
#> |:------------|-------:|-----:|---------:|------------:|------:|-----:|:-----:|:---------:|--------:|------:|-----:|
#> |Oranges      |       0|     1|         0|            0|      1|     2|  19   |     3     |        3|       |      |
#> |Pears        |       0|     0|         0|            0|      0|     0|   4   |     5     |        5|       |     x|
#> |Chocolate    |       1|     1|         0|            1|      1|     4|  48   |     1     |        1|      x|      |
#> |Strawberries |       1|     1|         0|            0|      1|     3|  28   |     2     |        2|       |      |
#> |Sweets       |       0|     1|         0|            0|      0|     1|  12   |     4     |        4|       |      |
#> 
#> SafeRank fuzz on Borda scores: 4.242641
#> Gaps in Borda scores: min 4 mean 9.6 max 20 ; all 4 7 8 9 20
#> Condorcet winner: Chocolate
#> Condorcet loser: Pears
#> 
#> 
#> Initial ranking found after 18 randomly-selected ballots were counted.
#> Final ranking:
#>      Oranges        Pears    Chocolate Strawberries       Sweets 
#>            3            5            1            2            4 
#> 
#> No subset of ballots produced the initial ranking