Count preferential ballots using an STV method

The votes parameter is as described in the Details for condorcet() with the following additional semantics. If a ballot fails to rank all candidates, it is informal (i.e. "invalid") unless invalid.partial=FALSE. If a ballot has equally-ranked candidates, it is informal unless equal.preferences=TRUE. Informal ballots are removed from the ballot box before counting, a warning is issued, and the ballots

Usage

stv(
  votes,
  nseats = NULL,
  eps = 0.001,
  equal.ranking = FALSE,
  fsep = "\t",
  ties = c("f", "b"),
  constant.quota = FALSE,
  quota.hare = FALSE,
  group.nseats = NULL,
  group.members = NULL,
  complete.ranking = FALSE,
  invalid.partial = FALSE,
  verbose = FALSE,
  seed = NULL,
  quiet = FALSE,
  digits = 3,
  backwards.compatible = FALSE,
  safety = 1,
  ...
)

Arguments

votes

an array with one column per candidate and one row per ballot, as described in the Details for condorcet()

nseats

the number of seats to be filled in this election

eps

fuzz-factor when comparing fractional votes

equal.ranking

= FALSE,

fsep

column-separator for output

ties

vector of tie-breaking methods: 'f' for forward, 'b' for backward

constant.quota

TRUE if quota is held constant. Over-rides quota.hare. Default is FALSE.

quota.hare

TRUE if Hare quota, FALSE if Droop quota (default)

group.nseats

number of seats reserved to members of a group, as in the Church of England's STV methodology.

group.members

vector of members of a group with reserved seats

complete.ranking

TRUE (default) if a complete ranking of all candidates will be produced. This parameter affects the value assigned to nseats when stv() is called with nseats=NULL, but has no effect on elections in which nseats is explicitly specified.

invalid.partial

TRUE if ballots which do not specify a complete ranking of candidates are informal (aka "invalid") i.e. ignored (with a warning). Default is FALSE.

verbose

TRUE for diagnostic output

seed

integer seed for tie-breaking. Warning: if non-NULL, the PRNG for R is reseeded prior to every random tie-break among the possibly-elected candidates. We have preserved this functionality in this branch to allow regression against the legacy codebase of vote_2.3.2. In SafeVote the default value for seed is NULL rather than the legacy value of 1234, to mitigate the validity hazard of PRNG reseedings during a stochastic experiment.

quiet

TRUE to suppress console output

digits

number of significant digits in the output table

backwards.compatible

TRUE to regress against vote2_3.2 by disabling $margins, $fuzz, $rankingTable, $safeRank

safety

number of standard deviations on vote-counts, when producing a safeRank by clustering near-ties in a complete ranking

...

undocumented intent (preserved from legacy code)

Value

object of class vote.stv. Note: the winning margins in this object are valid for the elected candidates and their (total) ranking, but must be adjusted within tiegroups to be valid for the candidates' (possibly partial) safeRank.

Examples

data(food_election)
stv(food_election, safety = 0.0)
#> 
#> Results of Single transferable vote
#> ===================================                           
#> Number of valid votes:   20
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          2
#> 
#> Fuzz on safeRank:	 0
#> Complete ranking:
#> 
#> |             | Rank|SafeRank |  Margin   |
#> |:------------|----:|:--------|:---------:|
#> |Oranges      |    2|2        | 1.4451111 |
#> |Pears        |    4|4        |    NA     |
#> |Chocolate    |    1|1        | 8.0000000 |
#> |Strawberries |    3|3        | 1.7774444 |
#> |Sweets       |    3|3        | 0.7774444 |
#> 
#> Vote transfers:
#> 
#> |             |         1| 2-trans|     2| 3-trans|      3| 4-trans|       4|
#> |:------------|---------:|-------:|-----:|-------:|------:|-------:|-------:|
#> |Oranges      |     4.000|   0.000| 4.000|       2|  6.000|   0.000|   6.000|
#> |Pears        |     2.000|   0.000| 2.000|      -2|       |        |        |
#> |Chocolate    |    12.000|  -5.332|      |        |       |        |        |
#> |Strawberries |     1.000|   3.555| 4.555|       0|  4.555|   0.000|   4.555|
#> |Sweets       |     1.000|   1.777| 2.777|       0|  2.777|  -2.777|        |
#> |Tie-breaks   |          |        |      |        |       |        |        |
#> |Elected      | Chocolate|        |      |        |       |        | Oranges|
#> |Eliminated   |          |        | Pears|        | Sweets|        |        |
#> 
#> Elected: Chocolate, Oranges 
#> 
stv(food_election, nseats = 2)
#> 
#> Results of Single transferable vote
#> ===================================                           
#> Number of valid votes:   20
#> Number of invalid votes:  0
#> Number of candidates:     5
#> Number of seats:          2
#> 
#> Fuzz on safeRank:	 4.472136
#> Complete ranking:
#> 
#> |             | Rank|SafeRank |  Margin   |
#> |:------------|----:|:--------|:---------:|
#> |Oranges      |    2|2        | 1.4451111 |
#> |Pears        |    4|2        |    NA     |
#> |Chocolate    |    1|1        | 8.0000000 |
#> |Strawberries |    3|2        | 1.7774444 |
#> |Sweets       |    3|2        | 0.7774444 |
#> 
#> Vote transfers:
#> 
#> |             |         1| 2-trans|     2| 3-trans|      3| 4-trans|       4|
#> |:------------|---------:|-------:|-----:|-------:|------:|-------:|-------:|
#> |Oranges      |     4.000|   0.000| 4.000|       2|  6.000|   0.000|   6.000|
#> |Pears        |     2.000|   0.000| 2.000|      -2|       |        |        |
#> |Chocolate    |    12.000|  -5.332|      |        |       |        |        |
#> |Strawberries |     1.000|   3.555| 4.555|       0|  4.555|   0.000|   4.555|
#> |Sweets       |     1.000|   1.777| 2.777|       0|  2.777|  -2.777|        |
#> |Tie-breaks   |          |        |      |        |       |        |        |
#> |Elected      | Chocolate|        |      |        |       |        | Oranges|
#> |Eliminated   |          |        | Pears|        | Sweets|        |        |
#> 
#> Elected: Chocolate, Oranges 
#>