Assess the safety of a preliminary result for an election

Ballots are deleted at random from the ballot-box, with election results computed once per dinc ballot-deletions. The experiment terminates after a specified number of ballots have been deleted, or a specified number of ballot-counts have occurred. Note: these ballot-counts are correlated. Use testFraction() to experiment with independently-drawn samples from the ballot-box.

Usage

testDeletions(
  votes,
  countMethod = "stv",
  countArgs = list(),
  dstart = NULL,
  dinc = NULL,
  dlimit = NULL,
  drep = NULL,
  rankMethod = "safeRank",
  exptName = NULL,
  equiet = FALSE,
  everbose = FALSE
)

Arguments

votes

A set of ballots, as in vote_2.3.2

countMethod

"stv" (default) or "condorcet"

countArgs

List of args to be passed to countMethod (in addition to votes)

dstart

Number of ballots in the first ballot-count (selected at random from votes, without replacement)

dinc

Number of ballots to be deleted in subsequent steps

dlimit

Maximum number of ballots to delete (in addition to dstart)

drep

Maximum number of elections (required if dinc=0)

rankMethod

"safeRank" (default), "elected", or "rank". "rank" is a total ranking of the candidates, with ties broken at random. "elected" assigns rank=1 to elected candidates, rank=2 for eliminated candidates.

exptName

stem-name of experimental units e.g. "E". If NULL, then a 3-character string of capital letters is chosen at random.

equiet

TRUE to suppress all experimental output

everbose

TRUE to produce diagnostic output from the experiment

Value

SafeRankExpt

object, describing this experiment and its results

Examples

data(food_election)
testDeletions(food_election)
#> Number of ballots counted by stv: 20, 18, 16, 14, 12, 10, 8, 6, 4, 2
#> 
#> Results of testDeletions at 2023-01-14 21:14:53
#> 
#> Dataset = food_election, countMethod = stv, rankMethod = safeRank
#> 
#> |             | dstart| dinc| dlimit| drep|
#> |:------------|------:|----:|------:|----:|
#> |otherFactors |     20|    2|      2|   10|
#> 
#> Unit factors: initSample, removedBallots
#> 
#> Experiment ID, number of ballots in simulated election, ranks, winning margins:
#> 
#> |exptID | nBallots| Oranges| Pears| Chocolate| Strawberries| Sweets| m.Oranges| m.Pears| m.Chocolate| m.Strawberries| m.Sweets|
#> |:------|--------:|-------:|-----:|---------:|------------:|------:|---------:|-------:|-----------:|--------------:|--------:|
#> |EPS0   |       20|       2|     2|         1|            2|      2|  6.000000|       0|           8|       4.554889| 2.777444|
#> |EPS1   |       18|       2|     2|         1|            2|      2|  6.000000|       0|           6|       3.799300| 2.199700|
#> |EPS2   |       16|       2|     2|         1|            2|      2|  5.000000|       0|           6|       3.443778| 2.221889|
#> |EPS3   |       14|       2|     2|         1|            2|      2|  4.000000|       0|           7|       2.888222| 2.444111|
#> |EPS4   |       12|       2|     2|         1|            2|      2|  3.500625|       2|           6|       2.000000| 0.000000|
#> |EPS5   |       10|       2|     2|         1|            2|      2|  0.000000|       2|           5|       2.094667| 1.571000|
#> |EPS6   |        8|       2|     2|         1|            2|      2|  0.000000|       2|           4|       1.666167| 1.666167|
#> |EPS7   |        6|       1|     1|         1|            1|      1|  0.000000|       2|           4|       0.999500| 0.999500|
#> |EPS8   |        4|       2|     2|         1|            2|      2|  0.000000|       0|           4|       1.332833| 1.332833|
#> |EPS9   |        2|       2|     2|         1|            2|      2|  0.000000|       0|           2|       1.332333| 0.000000|
testDeletions(food_election, countMethod="stv",
  countArgs=list(complete.ranking=TRUE))
#> Number of ballots counted by stv: 20, 18, 16, 14, 12, 10, 8, 6, 4, 2
#> 
#> Results of testDeletions at 2023-01-14 21:14:53
#> 
#> Dataset = food_election, countMethod = stv, rankMethod = safeRank
#> 
#> |          | complete.ranking|
#> |:---------|----------------:|
#> |countArgs |             TRUE|
#> 
#> 
#> |             | dstart| dinc| dlimit| drep|
#> |:------------|------:|----:|------:|----:|
#> |otherFactors |     20|    2|      2|   10|
#> 
#> Unit factors: initSample, removedBallots
#> 
#> Experiment ID, number of ballots in simulated election, ranks, winning margins:
#> 
#> |exptID | nBallots| Oranges| Pears| Chocolate| Strawberries| Sweets| m.Oranges| m.Pears| m.Chocolate| m.Strawberries|  m.Sweets|
#> |:------|--------:|-------:|-----:|---------:|------------:|------:|---------:|-------:|-----------:|--------------:|---------:|
#> |QUK0   |       20|       2|     2|         1|            2|      2|     4.000|       2|    8.000000|      6.7771111| 3.8885556|
#> |QUK1   |       18|       2|     2|         1|            2|      2|     4.000|       1|    7.000000|      6.8174545| 3.1815455|
#> |QUK2   |       16|       2|     2|         1|            2|      2|     4.000|       1|    5.000000|      6.6287407| 1.7035926|
#> |QUK3   |       14|       1|     1|         1|            1|      1|     4.000|       1|    7.999143|      4.9991429| 1.6665238|
#> |QUK4   |       12|       1|     1|         1|            1|      1|     4.000|       0|    6.332500|      4.3325000| 1.6665000|
#> |QUK5   |       10|       1|     1|         1|            1|      1|     4.000|       0|    5.000000|      2.6658667| 1.6664667|
#> |QUK6   |        8|       1|     1|         1|            1|      1|     3.000|       0|    5.000000|      2.9325333| 0.7331333|
#> |QUK7   |        6|       1|     1|         1|            1|      1|     2.000|       0|    4.249250|      2.2492500| 0.7497500|
#> |QUK8   |        4|       1|     1|         1|            1|      1|     2.000|       0|    2.000000|      0.6661667| 0.6661667|
#> |QUK9   |        2|       1|     1|         1|            1|      1|     1.799|       0|    1.799000|      0.7990000| 0.0000000|