Assess the safety of a preliminary result for an election

Ballots are deleted at random from the ballot-box, with election results computed once per dinc ballot-deletions. The experiment terminates after a specified number of ballots have been deleted, or a specified number of ballot-counts have occurred. Note: these ballot-counts are correlated. Use testFraction() to experiment with independently-drawn samples from the ballot-box.

Usage

testDeletions(
  votes,
  countMethod = "stv",
  countArgs = list(),
  dstart = NULL,
  dinc = NULL,
  dlimit = NULL,
  drep = NULL,
  rankMethod = "safeRank",
  exptName = NULL,
  quiet = FALSE,
  verbose = FALSE
)

Arguments

votes

A set of ballots, as in vote_2.3.2

countMethod

The name of a function which will count the ballots

countArgs

List of args to be passed to countMethod (in addition to votes)

dstart

Number of ballots in the first ballot-count (selected at random from votes, without replacement)

dinc

Number of ballots to be deleted in subsequent steps

dlimit

Maximum number of ballots to delete (in addition to dstart)

drep

Maximum number of elections (required if dinc=0)

rankMethod

Name of a ranking attribute in the output of countMethod

exptName

stem-name of experimental units e.g. "E". If NULL, then a 3-character string of capital letters is chosen at random.

quiet

TRUE to suppress all output

verbose

TRUE to produce diagnostic output

Value

SafeRankExpt object, describing this experiment and its results

Examples

data(food_election)
testDeletions(food_election)
#> Number of ballots counted by stv:
#>   20, 18, 16, 14, 12, 10, 8, 6, 4, 2
#> 
#> Results of testDeletions at 2022-12-27 05:56:25
#> 
#> Dataset = food_election, countMethod = stv, rankMethod = safeRank
#> 
#> |             | dstart| dinc| dlimit| drep|
#> |:------------|------:|----:|------:|----:|
#> |otherFactors |     20|    2|      2|   10|
#> 
#> Unit factors: initSample, removedBallots
#> 
#> Experiment ID, number of ballots in simulated election, ranks, winning margins:
#> 
#> |exptID | nBallots| Oranges| Pears| Chocolate| Strawberries| Sweets| m.Oranges|   m.Pears| m.Chocolate| m.Strawberries|  m.Sweets|
#> |:------|--------:|-------:|-----:|---------:|------------:|------:|---------:|---------:|-----------:|--------------:|---------:|
#> |ENB0   |       20|       2|     2|         1|            2|      2| 0.7774444| 1.7774444|           8|      2.0000000| 1.4451111|
#> |ENB1   |       18|       2|     2|         1|            2|      2| 1.3633636| 2.2722727|           7|      1.0000000| 0.3643636|
#> |ENB2   |       16|       2|     2|         1|            2|      2| 1.3997000| 0.2659667|           7|      1.0000000| 1.6003000|
#> |ENB3   |       14|       2|     2|         1|            2|      2| 1.2496250| 0.8330833|           5|      1.0000000| 0.9172917|
#> |ENB4   |       12|       1|     1|         1|            1|      1| 0.6663333| 0.6663333|           3|      1.0000000| 1.6673333|
#> |ENB5   |       10|       1|     1|         1|            1|      1| 0.6662667| 0.9994000|           2|      0.3337333| 2.0006000|
#> |ENB6   |        8|       1|     1|         1|            1|      1| 0.3330833| 0.9992500|           2|      0.6669167| 1.0007500|
#> |ENB7   |        6|       1|     1|         1|            1|      1| 0.0000000| 0.0000000|           1|      1.0000000| 1.0000000|
#> |ENB8   |        4|       1|     1|         1|            1|      1| 0.0000000| 0.0000000|           0|      1.0000000| 1.0000000|
#> |ENB9   |        2|       1|     2|         2|            2|      2| 2.0000000| 0.0000000|           0|      0.0000000| 0.0000000|
testDeletions(food_election, countMethod="stv",
  countArgs=list(complete.ranking=TRUE))
#> Number of ballots counted by stv:
#>   20, 18, 16, 14, 12, 10, 8, 6, 4, 2
#> 
#> Results of testDeletions at 2022-12-27 05:56:25
#> 
#> Dataset = food_election, countMethod = stv, rankMethod = safeRank
#> 
#> |          | complete.ranking|
#> |:---------|----------------:|
#> |countArgs |             TRUE|
#> 
#> 
#> |             | dstart| dinc| dlimit| drep|
#> |:------------|------:|----:|------:|----:|
#> |otherFactors |     20|    2|      2|   10|
#> 
#> Unit factors: initSample, removedBallots
#> 
#> Experiment ID, number of ballots in simulated election, ranks, winning margins:
#> 
#> |exptID | nBallots| Oranges| Pears| Chocolate| Strawberries| Sweets| m.Oranges|  m.Pears| m.Chocolate| m.Strawberries|  m.Sweets|
#> |:------|--------:|-------:|-----:|---------:|------------:|------:|---------:|--------:|-----------:|--------------:|---------:|
#> |MQC0   |       20|       2|     2|         1|            2|      2| 1.8885556| 2.777111|           8|      2.0000000| 0.1114444|
#> |MQC1   |       18|       2|     2|         1|            2|      2| 1.9996667| 2.999667|          10|      2.0000000| 0.0000000|
#> |MQC2   |       16|       2|     2|         1|            2|      2| 2.0299394| 2.272454|           9|      1.0000000| 1.9990000|
#> |MQC3   |       14|       3|     3|         1|            2|      3| 0.2997000| 1.000000|           8|      4.0662667| 1.9990000|
#> |MQC4   |       12|       2|     2|         1|            2|      2| 0.2496250| 2.499750|           6|      1.0000000| 1.9990000|
#> |MQC5   |       10|       2|     2|         1|            2|      2| 1.4441111| 1.888222|           4|      0.5558889| 0.4438611|
#> |MQC6   |        8|       2|     2|         1|            2|      2| 0.3338333| 1.199400|           3|      0.5337333| 1.1324333|
#> |MQC7   |        6|       2|     2|         1|            2|      2| 1.5996000| 1.799800|           4|      0.0000000| 0.0000000|
#> |MQC8   |        4|       2|     2|         1|            2|      2| 1.6661667| 0.000000|           4|      0.0000000| 0.0000000|
#> |MQC9   |        2|       2|     2|         1|            2|      2| 0.8328333| 0.000000|           2|      0.0000000| 0.0000000|