Assess the safety of a preliminary result for an election

Ballots are deleted at random from the ballot-box, with election results computed once per 'dinc' ballot-deletions. The experiment terminates after a specified number of ballots have been deleted, or a specified number of ballot-counts have occurred. Note: these ballot-counts are correlated. Use testFraction() to experiment with independently-drawn samples from the ballot-box.

Usage

testDeletions(
  votes = "food_election",
  countMethod = "condorcet",
  countArgs = NULL,
  dlimit = NULL,
  dstart = NULL,
  dinc = NULL,
  drep = NULL,
  rankMethod = "safeRank",
  exptName = NULL,
  quiet = FALSE,
  verbose = FALSE
)

Arguments

votes

A set of ballots

countMethod

The name of a function which will count the ballots

countArgs

List of args to be passed to countMethod (in addition to votes)

dlimit

Maximum number of ballots to delete

dstart

Number of ballots to be deleted after the initial ballot-count

dinc

Number of ballots to be deleted in subsequent steps

drep

Maximum number of elections (required if dinc=0)

rankMethod

Name of a ranking attribute in the output of countMethod

exptName

stem-name of experimental units e.g. "E"

quiet

TRUE to suppress all output

verbose

TRUE to produce diagnostic output

Value

SafeRankExpt object, describing this experiment and its results

Examples

data(food_election)
testDeletions(food_election)
#> Number of ballots counted by condorcet :
#>  20, 18, 16, 14, 12, 10, 8, 6, 4, 2
#> 
#> Experimental results:
#> $nBallots
#> PLC0 PLC1 PLC2 PLC3 PLC4 PLC5 PLC6 PLC7 PLC8 PLC9 
#>   20   18   16   14   12   10    8    6    4    2 
#> 
#> $ranks
#>      Oranges Pears Chocolate Strawberries Sweets
#> PLC0       2     2         1            2      2
#> PLC1       2     4         1            2      4
#> PLC2       2     2         1            2      2
#> PLC3       2     4         1            2      4
#> PLC4       2     4         1            2      4
#> PLC5       1     1         1            1      1
#> PLC6       1     1         1            1      1
#> PLC7       1     1         1            1      1
#> PLC8       1     1         3            3      3
#> PLC9       1     2         3            3      3
#> 
#> $margins
#>      Oranges Pears Chocolate Strawberries Sweets
#> PLC0       6    NA        20            6      8
#> PLC1      10    NA        19            3      4
#> PLC2       6    NA        18            4      4
#> PLC3       9    NA        18            9      1
#> PLC4       3     2        10            7     NA
#> PLC5       6     5         2            4     NA
#> PLC6       5     5         2            1     NA
#> PLC7       2     2         2            3     NA
#> PLC8       2     4         1           NA      3
#> PLC9       3     4        NA           NA     NA
#> 
#> $countMethod
#> [1] "condorcet"
#> 
#> $rankMethod
#> [1] "safeRank"
#> 
#> $datasetName
#> [1] "food_election"
#> 
#> $experimentalMethod
#> [1] "testDeletions"
#> 
#> $otherFactors
#> $otherFactors$dlimit
#> [1] 18
#> 
#> $otherFactors$dstart
#> [1] 0
#> 
#> $otherFactors$dinc
#> [1] 2
#> 
#> $otherFactors$drep
#> [1] 10
#> 
#> $otherFactors$countArgs
#> NULL
#> 
#> 
#> attr(,"class")
#> [1] "SafeRankExpt"
testDeletions(food_election, countMethod="stv",
  countArgs=list(complete.ranking=TRUE))
#> Number of ballots counted by stv :
#>  20, 18, 16, 14, 12, 10, 8, 6, 4, 2
#> 
#> Experimental results:
#> $nBallots
#> NHQ0 NHQ1 NHQ2 NHQ3 NHQ4 NHQ5 NHQ6 NHQ7 NHQ8 NHQ9 
#>   20   18   16   14   12   10    8    6    4    2 
#> 
#> $ranks
#>      Oranges Pears Chocolate Strawberries Sweets
#> NHQ0       2     2         1            2      2
#> NHQ1       2     2         1            2      2
#> NHQ2       2     2         1            2      2
#> NHQ3       1     1         1            1      1
#> NHQ4       1     1         1            1      1
#> NHQ5       1     1         1            1      1
#> NHQ6       1     1         1            1      1
#> NHQ7       1     1         1            1      1
#> NHQ8       1     1         1            1      1
#> NHQ9       1     2         2            2      2
#> 
#> $margins
#>         Oranges Pears Chocolate Strawberries    Sweets
#> NHQ0 0.11144444    NA         8    2.7771111 1.8885556
#> NHQ1 0.09127273    NA         7    1.0902727 1.9087273
#> NHQ2 0.88922222    NA         5    0.2215556 1.1107778
#> NHQ3 1.00042857    NA         3    0.6660952 0.3334762
#> NHQ4 1.33400000    NA         2    0.3330000 1.3330000
#> NHQ5 0.00000000    NA         3    0.3995000 0.6997500
#> NHQ6 0.00000000 0.000         2    0.9990000        NA
#> NHQ7 0.00000000 0.000         1    0.1990000        NA
#> NHQ8 1.00000000 0.401         0    0.5990000        NA
#> NHQ9 2.00000000 0.000         0    0.0000000        NA
#> 
#> $countMethod
#> [1] "stv"
#> 
#> $rankMethod
#> [1] "safeRank"
#> 
#> $datasetName
#> [1] "food_election"
#> 
#> $experimentalMethod
#> [1] "testDeletions"
#> 
#> $otherFactors
#> $otherFactors$dlimit
#> [1] 18
#> 
#> $otherFactors$dstart
#> [1] 0
#> 
#> $otherFactors$dinc
#> [1] 2
#> 
#> $otherFactors$drep
#> [1] 10
#> 
#> $otherFactors$countArgs
#> $otherFactors$countArgs$complete.ranking
#> [1] TRUE
#> 
#> 
#> 
#> attr(,"class")
#> [1] "SafeRankExpt"