Bootstrapping experiment, with fractional counts of a ballot box.

Starting from some number ('astart') of randomly-selected ballots, an increasingly-large collection of randomly-selected ballots are counted. The ballots are chosen independently without replacement for each experimental unit; if you want to count decreasingly-sized portions of a single sample of ballots, use testDeletions(). The rankings and margins produced by each simulated election are returned as a matrix with one row per election.

Usage

testFraction(
  votes,
  astart = NULL,
  ainc = NULL,
  arep = NULL,
  rankMethod = "safeRank",
  countMethod = "stv",
  countArgs = list(),
  exptName = NULL,
  quiet = FALSE,
  verbose = FALSE
)

Arguments

votes

A numeric matrix: one row per ballot, one column per candidate

astart

Starting number of ballots (min 2)

ainc

Number of ballots to be added in each step

arep

Limit on the number of repetitions of the test. Required to be non-null if ainc==0.

rankMethod

Name of a ranking attribute in the output of countMethod, e.g. "elected", "ranking", "safeRank".

countMethod

The name of a function which will count the ballots, e.g. "stv", "condorcet"

countArgs

List of args to be passed to countMethod (in addition to votes)

exptName

stem-name of experimental units e.g. "E"

quiet

TRUE to suppress all output

verbose

TRUE to produce diagnostic output

Value

a matrix of experimental results, of dimension n by 2m+1, where n is the number of elections and m is the number of candidates. The first column is named "nBallots". Other columns indicate the ranking and margin of each candidate in each election.

Examples

data(food_election)
testFraction(food_election, countMethod="condorcet",
             countArgs=list(safety=0.5,complete.ranking=TRUE))
#> Fraction of condorcet ballots:
#>  10%, 20%, 30%, 40%, 50%, 60%, 70%, 80%, 90%
#> 
#> Results of testFraction at 2022-12-08 22:22:18
#> 
#> Dataset = food_election, countMethod = condorcet, rankMethod = safeRank
#> 
#> |          | safety| complete.ranking|
#> |:---------|------:|----------------:|
#> |countArgs |    0.5|             TRUE|
#> 
#> 
#> |             | astart| ainc| arep|
#> |:------------|------:|----:|----:|
#> |otherFactors |      2|    2|    9|
#> 
#> Experiment ID, number of ballots in simulated election, ranks, winning margins:
#> 
#> |exptID | nBallots| Oranges| Pears| Chocolate| Strawberries| Sweets| m.Oranges| m.Pears| m.Chocolate| m.Strawberries| m.Sweets|
#> |:------|--------:|-------:|-----:|---------:|------------:|------:|---------:|-------:|-----------:|--------------:|--------:|
#> |VJE1   |        2|       3|     3|         1|            2|      3|          |        |           2|              6|         |
#> |VJE2   |        4|       3|     5|         1|            2|      3|         3|        |           6|              2|        1|
#> |VJE3   |        6|       2|     4|         1|            2|      4|         2|        |           7|              3|        1|
#> |VJE4   |        8|       4|     4|         1|            2|      3|          |        |          12|              7|        9|
#> |VJE5   |       10|       2|     5|         1|            2|      2|        10|        |          12|              2|        2|
#> |VJE6   |       12|       3|     4|         1|            2|      4|         9|       2|           9|              4|         |
#> |VJE7   |       14|       3|     5|         1|            2|      3|         9|        |          15|             13|        3|
#> |VJE8   |       16|       3|     5|         1|            2|      4|         5|        |          14|              4|        5|
#> |VJE9   |       18|       2|     5|         1|            2|      4|         6|        |          18|              6|        8|
testFraction(dublin_west, astart=20, ainc=20, arep=2, countMethod="stv",
             rankMethod="elected", quiet=FALSE)
#> Fraction of stv ballots:
#>  0.1%, 0.1%
#> 
#> Results of testFraction at 2022-12-08 22:22:18
#> 
#> Dataset = dublin_west, countMethod = stv, rankMethod = elected
#> 
#> |             | astart| ainc| arep|
#> |:------------|------:|----:|----:|
#> |otherFactors |     20|   20|    2|
#> 
#> Experiment ID, number of ballots in simulated election, ranks, winning margins:
#> 
#> |exptID | nBallots| Bonnie| Burton| Ryan| Higgins| Lenihan| McDonald| Morrissey| Smyth| Terry|  m.Bonnie|  m.Burton|    m.Ryan| m.Higgins| m.Lenihan| m.McDonald| m.Morrissey| m.Smyth|  m.Terry|
#> |:------|--------:|------:|------:|----:|-------:|-------:|--------:|---------:|-----:|-----:|---------:|---------:|---------:|---------:|---------:|----------:|-----------:|-------:|--------:|
#> |JXS1   |       20|      1|      2|    1|       2|       2|        2|         1|     1|     1| 0.6129017| 0.5715714| 0.2439554|         1|  3.428429|  0.5207901|   0.3776473|        | 1.050781|
#> |JXS2   |       40|      1|      2|    1|       2|       2|        1|         1|     1|     2| 0.2726364| 0.1371656| 0.8720582|         2|  2.409802|  1.7181399|   1.4098019|        | 2.727364|