Bootstrapping experiment, with fractional counts of a ballot box.

Starting from some number (astart) of randomly-selected ballots, an increasingly-large collection of randomly-selected ballots are counted. The ballots are chosen independently without replacement for each experimental unit; if you want to count decreasingly-sized portions of a single sample of ballots, use testDeletions().

Usage

testFraction(
  votes = NULL,
  astart = NULL,
  ainc = NULL,
  arep = NULL,
  trep = NULL,
  rankMethod = "safeRank",
  countMethod = "stv",
  countArgs = list(),
  exptName = NULL,
  quiet = FALSE,
  verbose = FALSE
)

Arguments

votes

A numeric matrix: one row per ballot, one column per candidate

astart

Starting number of ballots (min 2)

ainc

Number of ballots to be added in each step. Must be non-negative.

arep

Number of repetitions of the test on each step. Required to be non-NULL if ainc=0 && is.null(trep)`.

trep

Limit on the total number of simulated elections. Required to be non-NULL if ainc=0 && is.null(arep).

rankMethod

Name of a ranking attribute in the output of countMethod, e.g. "elected", "ranking", "safeRank".

countMethod

The name of a function which will count the ballots, e.g. "stv", "condorcet"

countArgs

List of args to be passed to countMethod (in addition to votes)

exptName

stem-name of experimental units e.g. "E". If NULL, then a 3-character string of capital letters is chosen at random.

quiet

TRUE to suppress all output

verbose

TRUE to produce diagnostic output

Value

a SafeRankExpt object of experimental results.

Examples

data(food_election)
testFraction(food_election, countMethod="condorcet",
             countArgs=list(safety=0.5,complete.ranking=TRUE))
#> Progress in counting condorcet ballots:
#>  10%, 20%,ADG2 elected 1 candidates, but ADG0 elected 0 candidates
#> Warning: 
#>  30%, 40%,ADG4 elected 1 candidates, but ADG0 elected 0 candidates
#> Warning: 
#>  50%,ADG5 elected 1 candidates, but ADG0 elected 0 candidates
#> Warning: 
#>  60%,ADG6 elected 1 candidates, but ADG0 elected 0 candidates
#> Warning: 
#>  70%,ADG7 elected 1 candidates, but ADG0 elected 0 candidates
#> Warning: 
#>  80%,ADG8 elected 1 candidates, but ADG0 elected 0 candidates
#> Warning: 
#>  90%,ADG9 elected 1 candidates, but ADG0 elected 0 candidates
#> Warning: 
#>  100%ADG10 elected 1 candidates, but ADG0 elected 0 candidates
#> Warning: 
#> 
#> 
#> Results of testFraction at 2022-12-22 16:09:04
#> 
#> Dataset = food_election, countMethod = condorcet, rankMethod = safeRank
#> 
#> |          | safety| complete.ranking|
#> |:---------|------:|----------------:|
#> |countArgs |    0.5|             TRUE|
#> 
#> 
#> |             | astart| ainc| arep|
#> |:------------|------:|----:|----:|
#> |otherFactors |      2|    2|    1|
#> 
#> Experiment ID, number of ballots in simulated election, ranks, winning margins:
#> 
#> |exptID | nBallots| Oranges| Pears| Chocolate| Strawberries| Sweets| m.Oranges| m.Pears| m.Chocolate| m.Strawberries| m.Sweets|
#> |:------|--------:|-------:|-----:|---------:|------------:|------:|---------:|-------:|-----------:|--------------:|--------:|
#> |ADG1   |        2|       1|     4|         1|            1|      4|         3|        |           1|              1|         |
#> |ADG2   |        4|       3|     5|         1|            3|      2|         3|        |           6|              1|        2|
#> |ADG3   |        6|       1|     3|         1|            3|      3|         1|       2|           5|               |        1|
#> |ADG4   |        8|       2|     4|         1|            2|      4|         1|        |           4|              7|         |
#> |ADG5   |       10|       4|     4|         1|            2|      2|          |       3|          15|              1|        5|
#> |ADG6   |       12|       2|     4|         1|            2|      4|         2|       1|          12|              5|         |
#> |ADG7   |       14|       3|     5|         1|            2|      4|         6|        |          14|              7|        5|
#> |ADG8   |       16|       3|     5|         1|            2|      4|         5|        |          14|              4|        5|
#> |ADG9   |       18|       3|     5|         1|            2|      3|        12|        |          24|              5|        3|
#> |ADG10  |       20|       3|     5|         1|            2|      4|         6|        |          20|              6|        8|
testFraction(dublin_west, astart=20, ainc=10, arep=2, trep=3, 
             countMethod="stv", rankMethod="elected", quiet=FALSE)
#> Progress in counting stv ballots:
#>  33.3%, 66.7%, 100%
#> 
#> Results of testFraction at 2022-12-22 16:09:04
#> 
#> Dataset = dublin_west, countMethod = stv, rankMethod = elected
#> 
#> |             | astart| ainc| arep|
#> |:------------|------:|----:|----:|
#> |otherFactors |     20|   10|    2|
#> 
#> Experiment ID, number of ballots in simulated election, ranks, winning margins:
#> 
#> |exptID | nBallots| Bonnie| Burton| Ryan| Higgins| Lenihan| McDonald| Morrissey| Smyth| Terry|  m.Bonnie|  m.Burton|    m.Ryan| m.Higgins| m.Lenihan| m.McDonald| m.Morrissey| m.Smyth|   m.Terry|
#> |:------|--------:|------:|------:|----:|-------:|-------:|--------:|---------:|-----:|-----:|---------:|---------:|---------:|---------:|---------:|----------:|-----------:|-------:|---------:|
#> |MLE1   |       20|      1|      2|    1|       2|       2|        1|         1|     1|     2| 1.0000000| 0.1022732| 1.5129044|  2.000000| 0.5645601|  0.0767145|   0.0767145|        | 1.0000000|
#> |MLE2   |       30|      1|      2|    2|       2|       2|        1|         1|     1|     1| 0.0000000| 0.4935210| 0.3998353|  1.000000| 3.6667778|  2.3328889|   1.6131226|        | 0.3868774|
#> |MLE3   |       20|      2|      1|    1|       2|       2|        2|         1|     1|     1| 0.0005417| 1.6664167| 0.6658333|  3.334167| 0.0000000|  0.0000417|   0.3335833|        | 0.0000000|